requirements.txt
# Core ML & NLP
torch>=2.1.0
transformers>=4.38.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Streamlit App
streamlit>=1.32.0

# Data Handling
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0
nltk>=3.8.1

# Utility & File I/O
scikit-learn>=1.3.0
jsonschema>=4.19.0

# Reranking & Generation
accelerate>=0.28.0
bitsandbytes>=0.43.0  # for GPU optimization (optional)
sentencepiece>=0.1.99
protobuf>=4.25.0

# Optional visualization & debugging
rich>=13.7.0
python-dotenv>=1.0.1
